"User who is not verify auth screen and close app and again opeen the app at that time ask for auth verification again. Step to produce :  Login with One Use -> access app for certain period of time and lock phone.  -> Now unlock Phone it will ask you to verify auth. -> Now Close App without validating user identity.  -> Again open the app -> it must ask for Auth verifcation"
"Let Assume You are logged in at 9:00 A.M auth token time is 2 Min and Refresh token time is 5 min. Also when ever new authtoken generated you got new refresh token time because of your last active time is change. So as per above time Your Auth token is generated at 9:02 A.M and refresh token is generated at 9:05 A.M

So if you access app for 3 mins so your auth token is updated at 9:02 A.M  Now due to this you new refresh token Expire Time is 9:07 A.M.

Now you close your app at 9:03 A.M and at 9:15 A.M you are trying to access your app. so as per rule your refresh token is expired so when you open app at 9:15 A.M at that time it will not ask you for Auth verification instead of this it will show you a Signin Screen

Please Note Down Highligthed statment"