Apk Link: Click Here to Download Patient and Provider Link Test Case Video: Click Here to Watch Testing Video
Test Cases Done Perform from Our End:
Note: During Testing we use 2 min as AUTH token Expired Time and 5 mins as Refresh token Expired Time from his last active time.
In short,
•	2 mins Ideal state for Auth Token
•	5 mins Ideal state for Refresh Token
Testing Devices:
•	MOTO G3: Android 6.0
•	MOTO G5s Plus Android 8.0
•	Realm 5 Pro Android 9.0
•
Case1: For Figure Print Module in Android
This case is tested for those devices which support figure print Authentication.
Scenario - 1: Login through valid credential and enable figure print during login. Now on successful login we keep app in ideal state didn’t perform any action inside Telehealer app for the 2 mins. So, after 2 mins we are clicking on app and it ask for auth token because it is expired after 2 mins. DONE
Scenario - 2: After successful login we have locked device and unlock it. On successful unlock it ask for Auth. That you will see in video which we share with you. In this video you see black screen which is the screen recorded view which display on lock. DONE
Scenario - 3:  In this scenario we put app in background after successful login. After 2 mins we put our Telehealer app in foreground. When app comes in foreground state it ask for auth because app is stay ideal for 2 mins. DONE
Scenario - 4: In this we have logged in app and kill app from the phone. Now app is not running in background for 2 mins. After 2 mins we have again lunch app and it ask for figure print because it is ideal for 2 mins and its authtoken in expired. DONE
Scenario - 5: Next three scenario is for refresh token. In this scenario if user is not performing any activity after login and app is in foreground for next 5 mins. And then after user try to perform action, at that time he will redirected to sign in in screen because he/she is ideal for 5 mins and due to this refresh token is expired. DONE
Scenario - 6: This scenario is also for refresh token and in that case if user logged in app and put app in background for next 5 mins he/she will be redirected to sign in screen because refreshed token is expired due to ideal state. DONE
Scenario - 7: In this scenario user logged in app and kills app. Now he will try to open app after 5 mins then he must need to login in sign in screen because refresh token is expired. DONE
Scenario - 8: In this scenario user logged in app and whenever authentication needed and user directly kill app and again try to open app at that time he/ she must verify his/her identity by providing valid fingered Print/ Pin / Password. DONE
Case2: For Pin Module in Android
 	Same as Case1 but use Pin
Case3: For Password Module in Android
 	Same as Case1 but use Password
•	If user has disable Quick login at that time it will ask for Password. Overall, 24 Scenario is Tested for Auth Module.

 Note:	Above All 8 scenario is working for Pin/ Password.
•	8 For Pin
•	8 For Figure Print
•	8 For Password